feat(component): ListeFormations: ajustement de la taille et du style des images vedettes
This commit adjusts the size and style of the featured images in the `ListeFormations` component. It introduces a new `imageVedetteSize` constant to define the width and height of the images, and applies these dimensions along with a `borderRadius` and `display: block` style to both the actual images and the skeleton placeholders. This ensures consistent sizing and appearance of the images throughout the component, improving the user experience.